[#52] Split release in its own subproject
