--- core/src/sbt-test/release/checkstyle-compile/project/plugins.sbt
@@ -1,1 +1,0 @@
-addSbtPlugin("com.nrinaudo" % "kantan.sbt" % sys.props("plugin.version"))
