--- core/src/sbt-test/release/checkstyle-compile/test
@@ -1,1 +1,0 @@
--> release
