--- core/src/sbt-test/release/checkstyle-test/test
@@ -1,1 +1,0 @@
--> release
