+++ release/src/sbt-test/release/checkstyle-main/build.sbt
@@ -1,5 +1,3 @@
-scalafmtVersion := "1.1.0"
-
 releaseProcess := Seq[ReleaseStep](
   kantan.sbt.release.KantanRelease.runCheckStyle
 )
