+++ release/src/sbt-test/release/checkstyle-main/test
@@ -1,0 +1,5 @@
+# Makes sure that ill-formatted files in src/main fail the release
+
+> test:checkStyle
+-> checkStyle
+-> release
