+++ release/src/sbt-test/release/checkstyle-test/build.sbt
@@ -1,5 +1,3 @@
-scalafmtVersion := "1.1.0"
-
 releaseProcess := Seq[ReleaseStep](
   kantan.sbt.release.KantanRelease.runCheckStyle
 )
