+++ release/src/sbt-test/release/checkstyle-test/project/plugins.sbt
@@ -1,0 +1,2 @@
+addSbtPlugin(
+  "com.nrinaudo" % "kantan.sbt-release" % sys.props("plugin.version"))
