+++ release/src/sbt-test/release/checkstyle-test/test
@@ -1,0 +1,5 @@
+# Makes sure that ill-formatted files in src/test fail the release
+
+> checkStyle
+-> test:checkStyle
+-> release
